pyproject.toml
[tool.poetry]
name = "stonks"
version = "0.1.0"
description = "Discord crypto price & chart bot"
authors = ["Christian Pérez Villanueva <perez.villanueva.christian34@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.19.2"
aiocoingecko = "^1.0.0"
requests = "^2.27.1"
lru-dict = "^1.1.7"
web3 = "^5.26.0"
pandas = "^1.3.5"
python-coinmarketcap = "^0.3"
html5lib = "^1.1"
beautifulsoup4 = "^4.10.0"
python-dateutil = "^2.8.2"
plotly = "^5.5.0"
inflection = "^0.5.1"
kaleido = [{ version = "<=0.2.1", platform = "win32" }, { version = "0.2.1.post1", platform = "linux" }]
tortoise-orm = {extras = ["asyncpg"], version = "^0.18.1"}
coinmarketcap_utils = {file = "coinmarketcap_utils/dist/coinmarketcap_utils-0.1.0.tar.gz"}
py-cord = {git = "https://github.com/Pycord-Development/pycord.git"}

[tool.poetry.dev-dependencies]
black = "^21.12b0"
pre-commit = "^2.18.1"
wemake-python-styleguide = "^0.16.1"
types-requests = "^2.27.16"
maturin = "^0.12.12"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
    "requests.*"
]

